feat: add common UI components
