+++ packages/components/README.md
@@ -1,0 +1,35 @@
+# Mock app components
+
+A collection of reusable React components for general-purpose use.
+
+## Installation
+
+### Install dependencies
+
+Make sure your machine has Node.js version 18 or later installed.
+
+To install the project dependencies, run the following command:
+
+```bash
+pnpm install
+```
+
+## Storybook
+
+Explore and interact with the components using Storybook.
+
+### Run Storybook
+
+```bash
+npm run storybook
+```
+
+Visit http://localhost:6006/ in your browser to view Storybook.
+
+## Unit Testing
+
+Run unit tests using Jest.
+
+```bash
+npm test
+```
