+++ packages/components/src/index.ts
@@ -3,5 +3,3 @@ export * from './components/Footer/Footer.js';
 export * from './components/Header/Header.js';
 export * from './components/Layout/Layout.js';
 export * from './components/ToastMessage/ToastMessage.js';
-export * from './utils/toastMessage.js'
-export * from './models/common.js'
