--- packages/components/src/models/common.ts
@@ -1,6 +1,0 @@
-export enum Status {
-  success = 'success',
-  error = 'error',
-  warning = 'warning',
-  info = 'info',
-}
